FIREBASE IS WORKING
