+++ controllers/api/trail-routes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { download } = require('express/lib/response');
-// const firebase = require('firebase');
+const firebase = require('firebase');
 const multer  = require('multer');
 const upload = multer();
 const { Trail, User, Comment, Rating } = require('../../models');
@@ -12,35 +12,37 @@ const withAuth = require('../../utils/auth');
 // gs://treaddit.appspot.com/trails/project3.jpg
 
 
-// const firebaseConfig = {
-//     apiKey: "AIzaSyDyyFmd6Y7okq8KMn7JyROKxfk46gKJfC4",
-//     authDomain: "treaddit.firebaseapp.com",
-//     projectId: "treaddit",
-//     storageBucket: "treaddit.appspot.com",
-//     messagingSenderId: "964574079370",
-//     appId: "1:964574079370:web:7dd35ffdb6443410a78073",
-//     measurementId: "G-60MZQ7M7LG"
-// };
+const firebaseConfig = {
+    apiKey: "AIzaSyDyyFmd6Y7okq8KMn7JyROKxfk46gKJfC4",
+    authDomain: "treaddit.firebaseapp.com",
+    projectId: "treaddit",
+    storageBucket: "treaddit.appspot.com",
+    messagingSenderId: "964574079370",
+    appId: "1:964574079370:web:7dd35ffdb6443410a78073",
+    measurementId: "G-60MZQ7M7LG"
+};
 
-// firebase.initializeApp(firebaseConfig);
+firebase.initializeApp(firebaseConfig);
 
-// var storage = firebase.storage();
-// var storageRef = storage.ref();
+var storage = firebase.storage();
+var storageRef = storage.ref();
 
-// const uploadTrailImageFirebase = (trailImage) => {
-//     var imageRef = storageRef.child(`/trails/${trailImage.originalname}`)
-//     var metaData = {
-//         contentType: 'image/jpeg'
-//     }
-//     return imageRef.put(trailImage.buffer, metaData).then((snapshot) => {
-//         console.log('Uploaded a blob or file!');
-//         console.log(snapshot);
+const uploadTrailImageFirebase = (trailImage) => {
+    var imageRef = storageRef.child(`/trails/${trailImage.originalname}`)
+    var metaData = {
+        contentType: 'image/jpeg'
+    }
+    return imageRef.put(trailImage.buffer, metaData).then((snapshot) => {
+        console.log('Uploaded a blob or file!');
+        console.log(snapshot);
 
-//         // snapshot._delegate.metadata.fullPath is to store in Trails DB
+        // snapshot._delegate.metadata.fullPath is to store in Trails DB
 
 
-//       });
-// }
+      });
+}
+
+
 
 function downloadTrailImage(img_ref) {
   
@@ -219,14 +221,16 @@ router.get('/:id', (req, res) => {
     });
 });
 
-router.post('/', (req, res) => {
-    console.log('req.file:', req.file)
-    console.log('req.body:', req.body)
-    // uploadTrailImageFirebase(req.file)
-    // .then((result) => {
-    //     // res.json(result);
-    // console.log('res.json(result):', res.json(result))
-
+router.post('/', upload.single("file"), (req, res) => {
+    // console.log('req.file:', req.file)
+    // console.log('req.body:', req.body)
+    uploadTrailImageFirebase(req.file)
+    .then((result) => {
+        res.json(result);
+        var img_ref = "/trails/" + req.file.originalname;    
+        return downloadTrailImage(img_ref)
+    })
+    .then(url => {
     Trail.create({
         name: req.body.name,
         length: req.body.length,
@@ -234,14 +238,14 @@ router.post('/', (req, res) => {
         bike_friendly: req.body.bike_friendly,
         difficulty: req.body.difficulty,
         description: req.body.description,
-        // img_ref: "img/" + req.body.name 
+        img_ref: url
     })
-    .then(dbTrailData => res.json(dbTrailData))
+    // .then(dbTrailData => res.json(dbTrailData))
     .catch(err => {
         console.log(err);
-        res.status(500).json(err);
+        // res.status(500).json(err);
     });
-});
+})});
 
 
 router.post('/rating/:id', withAuth, (req, res) => {
