+++ views/homepage.handlebars
@@ -36,7 +36,7 @@
                     </div>
                 </nav>
                     <a href="#demo{{id}}" data-toggle="collapse">
-                        <img src="/img/example.jpg" class="img-circle card" alt="Trail Image" width="255" height="255">
+                        <img src="{{img_ref}}" class="img-circle card" alt="Trail Image" width="255" height="255">
                     </a><br>
                     <ul class="list-inline">
                         <li><span class="glyphicon glyphicon-star"></span></li>
